Changing positioning
Because this can be called on any element, at any time, we don’t want
to rely on the position of the mouse.

This change is to align the center of each particle to the center of
the element that is being celebrated.

This also allows us to remove the mouseX & mouseY variables, which I
think also looks a little cleaner.